+++ tests/lib/testem/saucie-disconnect.js
@@ -1,0 +1,8 @@
+#!/usr/bin/env node
+
+"use strict";
+
+var saucie = require("saucie");
+var pidFile = "/tmp/sc_client.pid";
+
+saucie.disconnect(pidFile);