+++ ReleaseNotes.md
@@ -115,6 +115,7 @@ categorized by license:
 * [Normalize v3.0.1](https://necolas.github.io/normalize.css/)
 * [Buzz v1.1.0](http://buzz.jaysalvat.com)
 * [html5shiv v3.7.2](https://code.google.com/p/html5shiv/)
+* [Testem Sauce Labs tunnel scripts](https://github.com/testem/testem/tree/master/examples/saucelabs)
 
 ### zlib/libpng License ###
 * [fastXmlPull is based on XML for Script's Fast Pull Parser v3.1](http://wiki.fluidproject.org/display/fluid/Licensing+for+fastXmlPull.js)